pyproject.toml
[project]
name = "banking-tools-mcp"
version = "1.0.0"
description = "Basic Banking Tools MCP Server"
requires-python = ">=3.8"
dependencies = [
    "uipath-mcp>=0.0.101",
    "fastapi",
    "sqlalchemy[asyncpg]",
    "pandas",
    "python-dotenv",
    "aiohttp",
    "uvicorn"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "*.py",
    "*.json",
    "*.md",
    "requirements.txt"
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["banking_tools"]
known-third-party = ["fastapi", "sqlalchemy", "pandas", "aiohttp"]
